Refactor Espoo BI to support a mock client
ESPi is disabling their API in the staging environment, but we'd like to
keep testing our jobs to reveal any possible regressions before
production.

By adding some extra indirection, we can have the BI jobs running in all
environments but just disable the HTTP client and exhaust the CSV stream
instead.
