+++ service/src/main/kotlin/fi/espoo/evaka/espoo/bi/EspooBiHttpClient.kt
@@ -1,0 +1,41 @@
+// SPDX-FileCopyrightText: 2017-2023 City of Espoo
+//
+// SPDX-License-Identifier: LGPL-2.1-or-later
+
+package fi.espoo.evaka.espoo.bi
+
+import com.github.kittinunf.fuel.core.FuelManager
+import com.github.kittinunf.fuel.core.extensions.authentication
+import com.github.kittinunf.fuel.core.requests.DefaultBody
+import fi.espoo.evaka.EspooBiEnv
+import fi.espoo.voltti.logging.loggers.error
+import mu.KotlinLogging
+
+class EspooBiHttpClient(
+    private val fuel: FuelManager,
+    private val env: EspooBiEnv,
+) : EspooBiClient {
+    private val logger = KotlinLogging.logger {}
+
+    override fun sendBiCsvFile(fileName: String, stream: EspooBiJob.CsvInputStream) {
+        logger.info("Sending BI CSV file $fileName")
+        val (_, _, result) =
+            fuel
+                .put("${env.url}/report", listOf("filename" to fileName))
+                .header("Content-type", "text/csv")
+                .authentication()
+                .basic(env.username, env.password.value)
+                .body(DefaultBody({ stream }))
+                .responseString()
+        result.fold(
+            { logger.info("Sent BI CSV file $fileName successfully (${stream.totalBytes} bytes)") },
+            { error ->
+                val meta = mapOf("errorMessage" to error.errorData.decodeToString())
+                logger.error(error, meta) {
+                    "Failed to send BI CSV file $fileName (${stream.totalBytes} bytes sent before failure)"
+                }
+                throw error
+            }
+        )
+    }
+}
